requirements.txt
boto3==1.35.33
Flask==3.0.3
Flask_Cors==5.0.0
protobuf==3.20.3
python-dotenv==1.0.1
gunicorn==22.0.0
google-cloud-dialogflow==2.33.0
spacy==3.8.2
en_core_web_sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.8.0/en_core_web_sm-3.8.0-py3-none-any.whl#sha256=1932429db727d4bff3deed6b34cfc05df17794f4a52eeb26cf8928f7c1a0fb85
pymongo==4.10.1